add support for non-gt multis
