+++ src/main/java/blockrenderer6343/BlockRenderer6343.java
@@ -1,25 +1,26 @@
 package blockrenderer6343;
 
+import java.util.Map;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.SidedProxy;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import cpw.mods.fml.common.event.FMLServerAboutToStartEvent;
-import cpw.mods.fml.common.event.FMLServerStartedEvent;
-import cpw.mods.fml.common.event.FMLServerStartingEvent;
-import cpw.mods.fml.common.event.FMLServerStoppedEvent;
-import cpw.mods.fml.common.event.FMLServerStoppingEvent;
+import cpw.mods.fml.common.network.NetworkCheckHandler;
+import cpw.mods.fml.relauncher.Side;
 
 @Mod(
         modid = BlockRenderer6343.MOD_ID,
         version = Tags.VERSION,
         name = BlockRenderer6343.MOD_NAME,
         acceptedMinecraftVersions = "[1.7.10]",
-        dependencies = " required-after:CodeChickenLib;" + " required-after:NotEnoughItems;")
+        dependencies = " required-after:CodeChickenLib;" + " required-after:NotEnoughItems;"
+                + "required-after:structurelib")
 public class BlockRenderer6343 {
 
     public static final String MOD_ID = "blockrenderer6343";
@@ -29,11 +30,16 @@ public class BlockRenderer6343 {
     @SidedProxy(clientSide = MOD_ID + ".ClientProxy", serverSide = MOD_ID + ".CommonProxy")
     public static CommonProxy proxy;
 
+    public static boolean isGTLoaded;
+    public static boolean isBartworksLoaded;
+
     @Mod.EventHandler
     // preInit "Run before anything else. Read your config, create blocks, items,
     // etc, and register them with the GameRegistry."
     public void preInit(FMLPreInitializationEvent event) {
         proxy.preInit(event);
+        isGTLoaded = Loader.isModLoaded("gregtech");
+        isBartworksLoaded = Loader.isModLoaded("bartworks");
     }
 
     @Mod.EventHandler
@@ -46,32 +52,6 @@ public class BlockRenderer6343 {
     // postInit "Handle interaction with other mods, complete your setup based on this."
     public void postInit(FMLPostInitializationEvent event) {
         proxy.postInit(event);
-    }
-
-    @Mod.EventHandler
-    public void serverAboutToStart(FMLServerAboutToStartEvent event) {
-        proxy.serverAboutToStart(event);
-    }
-
-    @Mod.EventHandler
-    // register server commands in this event handler
-    public void serverStarting(FMLServerStartingEvent event) {
-        proxy.serverStarting(event);
-    }
-
-    @Mod.EventHandler
-    public void serverStarted(FMLServerStartedEvent event) {
-        proxy.serverStarted(event);
-    }
-
-    @Mod.EventHandler
-    public void serverStopping(FMLServerStoppingEvent event) {
-        proxy.serverStopping(event);
-    }
-
-    @Mod.EventHandler
-    public void serverStopped(FMLServerStoppedEvent event) {
-        proxy.serverStopped(event);
     }
 
     public static void debug(String message) {
@@ -88,5 +68,10 @@ public class BlockRenderer6343 {
 
     public static void error(String message) {
         LOG.error(message);
+    }
+
+    @NetworkCheckHandler
+    public final boolean networkCheck(Map<String, String> remoteVersions, Side side) {
+        return true;
     }
 }
