+++ src/main/java/blockrenderer6343/CommonProxy.java
@@ -1,19 +1,12 @@
 package blockrenderer6343;
 
-import net.minecraft.block.Block;
-
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import cpw.mods.fml.common.event.FMLServerAboutToStartEvent;
-import cpw.mods.fml.common.event.FMLServerStartedEvent;
-import cpw.mods.fml.common.event.FMLServerStartingEvent;
-import cpw.mods.fml.common.event.FMLServerStoppedEvent;
-import cpw.mods.fml.common.event.FMLServerStoppingEvent;
 
 public class CommonProxy {
 
-    public static Block testBlock;
+    // Not needed serverside
 
     // preInit "Run before anything else. Read your config, create blocks, items,
     // etc, and register them with the GameRegistry."
@@ -25,14 +18,4 @@ public class CommonProxy {
     // postInit "Handle interaction with other mods, complete your setup based on this."
     public void postInit(FMLPostInitializationEvent event) {}
 
-    public void serverAboutToStart(FMLServerAboutToStartEvent event) {}
-
-    // register server commands in this event handler
-    public void serverStarting(FMLServerStartingEvent event) {}
-
-    public void serverStarted(FMLServerStartedEvent event) {}
-
-    public void serverStopping(FMLServerStoppingEvent event) {}
-
-    public void serverStopped(FMLServerStoppedEvent event) {}
 }
