+++ src/main/java/blockrenderer6343/integration/gregtech/GT_GUI_MultiblocksHandler.java
@@ -1,50 +1,24 @@
 package blockrenderer6343.integration.gregtech;
 
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Set;
 import java.util.UUID;
-import java.util.function.Consumer;
-import java.util.function.Predicate;
 
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.resources.I18n;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.MovingObjectPosition;
-import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
 
-import org.jetbrains.annotations.NotNull;
-
-import com.github.vfyjxf.nee.network.NEENetworkHandler;
-import com.github.vfyjxf.nee.network.packet.PacketNEIPatternRecipe;
-import com.gtnewhorizon.structurelib.StructureEvent;
 import com.gtnewhorizon.structurelib.StructureLibAPI;
-import com.gtnewhorizon.structurelib.alignment.constructable.ConstructableUtility;
 import com.gtnewhorizon.structurelib.alignment.constructable.IConstructable;
 import com.gtnewhorizon.structurelib.alignment.constructable.IConstructableProvider;
 import com.gtnewhorizon.structurelib.alignment.constructable.ISurvivalConstructable;
-import com.gtnewhorizon.structurelib.structure.AutoPlaceEnvironment;
-import com.gtnewhorizon.structurelib.structure.IStructureElement;
 import com.gtnewhorizon.structurelib.structure.ISurvivalBuildEnvironment;
 import com.mojang.authlib.GameProfile;
 
 import blockrenderer6343.BlockRenderer6343;
-import blockrenderer6343.api.utils.BlockPosition;
 import blockrenderer6343.api.utils.CreativeItemSource;
-import blockrenderer6343.api.utils.PositionedIStructureElement;
-import blockrenderer6343.client.utils.GuiText;
 import blockrenderer6343.client.world.ClientFakePlayer;
 import blockrenderer6343.common.GUI_MultiblocksHandler;
-import codechicken.lib.math.MathHelper;
-import codechicken.nei.recipe.GuiRecipe;
-import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 import gregtech.api.interfaces.metatileentity.IMetaTileEntity;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.interfaces.tileentity.ITurnable;
@@ -52,174 +26,13 @@ import gregtech.api.threads.GT_Runnable_MachineBlockUpdate;
 
 public class GT_GUI_MultiblocksHandler extends GUI_MultiblocksHandler<IConstructable> {
 
-    protected static final int TIER_BUTTON_X = LAYER_BUTTON_X;
-    protected static final int TIER_BUTTON_Y = LAYER_BUTTON_Y - ICON_SIZE_Y;
-    protected static final int PROJECT_BUTTON_X = 145;
-    protected static final int PROJECT_BUTTON_Y = -5;
-    protected static final int OVERLAY_BUTTON_X = 145 - ICON_SIZE_X - 5;
-    protected static final int OVERLAY_BUTTON_Y = -5;
-    private static final BlockPosition MB_PLACE_POS = new BlockPosition(0, 64, 0);
-    public static final int MAX_PLACE_ROUNDS = 2000;
-
-    protected static int tierIndex = 1;
-
-    protected static String guiTextTier;
-    protected static String guiTierButtonTitle;
-    protected static int initialTierButtonTitleWidth;
-    protected ClearGuiButton previousTierButton, nextTierButton;
-
-    private static EntityPlayer fakeMultiblockBuilder;
-
-    private final List<List<ItemStack>> candidates = new ArrayList<>();
-    protected List<PositionedIStructureElement> structureElements = new ArrayList<>();
-    protected Consumer<List<List<ItemStack>>> onCandidateChanged;
-
     public GT_GUI_MultiblocksHandler() {
         super();
-
-        previousTierButton = new ClearGuiButton(0, TIER_BUTTON_X, TIER_BUTTON_Y, ICON_SIZE_X, ICON_SIZE_Y, "<");
-        nextTierButton = new ClearGuiButton(
-                0,
-                TIER_BUTTON_X + ICON_SIZE_X,
-                TIER_BUTTON_Y,
-                ICON_SIZE_X,
-                ICON_SIZE_Y,
-                ">");
-        GuiButton projectMultiblocksButton = new GuiButton(
-                0,
-                PROJECT_BUTTON_X,
-                PROJECT_BUTTON_Y,
-                ICON_SIZE_X,
-                ICON_SIZE_Y,
-                "P");
-        GuiButton overlayMultiblocksButton = new GuiButton(
-                0,
-                OVERLAY_BUTTON_X,
-                OVERLAY_BUTTON_Y,
-                ICON_SIZE_X,
-                ICON_SIZE_Y,
-                "?");
-
-        buttons.put(previousTierButton, this::togglePreviousTier);
-        buttons.put(nextTierButton, this::toggleNextTier);
-        buttons.put(projectMultiblocksButton, this::projectMultiblocks);
-        buttons.put(overlayMultiblocksButton, this::neiOverlay);
-    }
-
-    @Override
-    protected void setLocalizationAndColor() {
-        super.setLocalizationAndColor();
-        guiTextTier = GuiText.Tier.getLocal();
-        previousTierButton.setColors(super.buttonColorEnabled, super.buttonColorDisabled, super.buttonColorHovered);
-        nextTierButton.setColors(super.buttonColorEnabled, super.buttonColorDisabled, super.buttonColorHovered);
-
-        guiTierButtonTitle = getTierButtonTitle();
-
-        FontRenderer fontRenderer = Minecraft.getMinecraft().fontRenderer;
-        initialTierButtonTitleWidth = fontRenderer.getStringWidth(guiTierButtonTitle);
-        nextTierButton.xPosition = TIER_BUTTON_X + ICON_SIZE_X
-                + initialTierButtonTitleWidth
-                - fontRenderer.getStringWidth("<") / 2;
-    }
-
-    public void setOnCandidateChanged(Consumer<List<List<ItemStack>>> callback) {
-        onCandidateChanged = callback;
-    }
-
-    @Override
-    protected void loadNewMultiblock() {
-        tierIndex = 1;
-        layerIndex = -1;
-        initializeSceneRenderer(true);
-        lastRenderingController = renderingController;
-    }
-
-    private void projectMultiblocks() {
-        EntityPlayer player = Minecraft.getMinecraft().thePlayer;
-        World baseWorld = Minecraft.getMinecraft().theWorld;
-        MovingObjectPosition lookingPos = player.rayTrace(10, 1);
-        if (lookingPos.typeOfHit == MovingObjectPosition.MovingObjectType.MISS) return;
-        int playerDir = MathHelper.floor_double((player.rotationYaw * 4F) / 360F + 0.5D) & 3;
-        ItemStack itemStack = stackForm.copy();
-        if (!baseWorld.isAirBlock(lookingPos.blockX, lookingPos.blockY + 1, lookingPos.blockZ)) return;
-        itemStack.getItem().onItemUse(
-                itemStack,
-                player,
-                baseWorld,
-                lookingPos.blockX,
-                lookingPos.blockY + 1,
-                lookingPos.blockZ,
-                0,
-                lookingPos.blockX,
-                lookingPos.blockY,
-                lookingPos.blockZ);
-        ConstructableUtility.handle(
-                getTriggerStack(),
-                player,
-                baseWorld,
-                lookingPos.blockX,
-                lookingPos.blockY + 1,
-                lookingPos.blockZ,
-                playerDir);
-        baseWorld.setBlockToAir(lookingPos.blockX, lookingPos.blockY + 1, lookingPos.blockZ);
-        baseWorld.removeTileEntity(lookingPos.blockX, lookingPos.blockY + 1, lookingPos.blockZ);
-    }
-
-    private void neiOverlay() {
-        NBTTagCompound recipeInputs = new NBTTagCompound();
-        GuiRecipe currentScreen = (GuiRecipe) Minecraft.getMinecraft().currentScreen;
-        Minecraft.getMinecraft().displayGuiScreen(currentScreen.firstGui);
-        for (int i = 0; i < ingredients.size(); i++) {
-            ItemStack itemStack = ingredients.get(i);
-            if (itemStack != null) {
-                NBTTagCompound itemStackNBT = new NBTTagCompound();
-                itemStack.writeToNBT(itemStackNBT);
-                itemStackNBT.setInteger("Count", itemStack.stackSize);
-                recipeInputs.setTag("#" + i, itemStackNBT);
-            }
-        }
-        NEENetworkHandler.getInstance().sendToServer(new PacketNEIPatternRecipe(recipeInputs, new NBTTagCompound()));
-    }
-
-    @NotNull
-    private static ItemStack getTriggerStack() {
-        return new ItemStack(StructureLibAPI.getDefaultHologramItem(), tierIndex);
-    }
-
-    private void toggleNextTier() {
-        tierIndex++;
-        guiTierButtonTitle = getTierButtonTitle();
-        initializeSceneRenderer(false);
-    }
-
-    private void togglePreviousTier() {
-        if (tierIndex > 1) {
-            tierIndex--;
-            guiTierButtonTitle = getTierButtonTitle();
-            initializeSceneRenderer(false);
-        }
     }
 
     @Override
     protected String getMultiblockName() {
         return I18n.format(stackForm.getDisplayName());
-    }
-
-    protected String getTierButtonTitle() {
-        return guiTextTier + ": " + tierIndex;
-    }
-
-    @Override
-    protected void drawButtonsTitle() {
-        super.drawButtonsTitle();
-
-        FontRenderer fontRenderer = Minecraft.getMinecraft().fontRenderer;
-        fontRenderer.drawString(
-                guiTierButtonTitle,
-                TIER_BUTTON_X + ICON_SIZE_X
-                        + (initialTierButtonTitleWidth - fontRenderer.getStringWidth(guiTierButtonTitle)) / 2,
-                TIER_BUTTON_Y + 2,
-                super.guiColorFont);
     }
 
     @Override
@@ -233,10 +46,9 @@ public class GT_GUI_MultiblocksHandler extends GUI_MultiblocksHandler<IConstruct
         renderer.world.unloadEntities(Arrays.asList(fakeMultiblockBuilder));
 
         IConstructable constructable = null;
-
-        int oStackSize = stackForm.stackSize;
-        stackForm.getItem().onItemUse(
-                stackForm,
+        ItemStack copy = stackForm.copy();
+        copy.getItem().onItemUse(
+                copy,
                 fakeMultiblockBuilder,
                 renderer.world,
                 MB_PLACE_POS.x,
@@ -246,7 +58,6 @@ public class GT_GUI_MultiblocksHandler extends GUI_MultiblocksHandler<IConstruct
                 MB_PLACE_POS.x,
                 MB_PLACE_POS.y,
                 MB_PLACE_POS.z);
-        stackForm.stackSize = oStackSize;
 
         TileEntity tTileEntity = renderer.world.getTileEntity(MB_PLACE_POS.x, MB_PLACE_POS.y, MB_PLACE_POS.z);
         ((ITurnable) tTileEntity).setFrontFacing(ForgeDirection.SOUTH);
@@ -277,80 +88,5 @@ public class GT_GUI_MultiblocksHandler extends GUI_MultiblocksHandler<IConstruct
 
         if (!GT_Runnable_MachineBlockUpdate.isCurrentThreadEnabled())
             GT_Runnable_MachineBlockUpdate.setCurrentThreadEnabled(true);
-    }
-
-    private void scanCandidates() {
-        candidates.clear();
-        if (selectedBlock != null) {
-            for (PositionedIStructureElement structureElement : structureElements) {
-                if (structureElement.x == selectedBlock.x && structureElement.y == selectedBlock.y
-                        && structureElement.z == selectedBlock.z) {
-
-                    IStructureElement.BlocksToPlace blocksToPlace = structureElement.element.getBlocksToPlace(
-                            renderingController,
-                            renderer.world,
-                            selectedBlock.x,
-                            selectedBlock.y,
-                            selectedBlock.z,
-                            getTriggerStack(),
-                            AutoPlaceEnvironment.fromLegacy(
-                                    CreativeItemSource.instance,
-                                    fakeMultiblockBuilder,
-                                    iChatComponent -> {}));
-                    if (blocksToPlace != null) {
-                        Predicate<ItemStack> predicate = blocksToPlace.getPredicate();
-                        Set<ItemStack> rawCandidates = CreativeItemSource.instance
-                                .takeEverythingMatches(predicate, false, 0).keySet();
-
-                        List<List<ItemStack>> stackedCandidates = new ArrayList<>();
-                        for (ItemStack rawCandidate : rawCandidates) {
-                            boolean added = false;
-                            for (List<ItemStack> stackedCandidate : stackedCandidates) {
-                                List<String> firstCandidateTooltip = stackedCandidate.get(0)
-                                        .getTooltip(fakeMultiblockBuilder, false);
-                                List<String> rawCandidateTooltip = rawCandidate
-                                        .getTooltip(fakeMultiblockBuilder, false);
-                                if (firstCandidateTooltip.size() > 1 && rawCandidateTooltip.size() > 1
-                                        && firstCandidateTooltip.get(1).equals(rawCandidateTooltip.get(1))) {
-                                    stackedCandidate.add(rawCandidate);
-                                    added = true;
-                                    break;
-                                }
-                            }
-                            if (!added) {
-                                List<ItemStack> newStackedCandidate = new ArrayList<>();
-                                newStackedCandidate.add(rawCandidate);
-                                stackedCandidates.add(newStackedCandidate);
-                            }
-                        }
-
-                        candidates.addAll(stackedCandidates);
-
-                        if (onCandidateChanged != null) {
-                            onCandidateChanged.accept(candidates);
-                        }
-                    }
-                    return;
-                }
-            }
-        }
-        if (onCandidateChanged != null) {
-            onCandidateChanged.accept(candidates);
-        }
-    }
-
-    @Override
-    protected void onBlockSelected() {
-        scanCandidates();
-    }
-
-    @SubscribeEvent
-    public void OnStructureEvent(StructureEvent.StructureElementVisitedEvent event) {
-        structureElements.add(
-                new PositionedIStructureElement(
-                        event.getX(),
-                        event.getY(),
-                        event.getZ(),
-                        (IStructureElement<IConstructable>) event.getElement()));
     }
 }
