+++ src/main/resources/assets/blockrenderer6343/lang/en_US.lang
@@ -5,3 +5,8 @@ gui.blockrenderer6343.BgColor=C6C6C6
 gui.blockrenderer6343.ButtonHoveredColor=FFFFA0
 gui.blockrenderer6343.ButtonEnabledColor=202020
 gui.blockrenderer6343.ButtonDisabledColor=A0A0A0
+
+blockrenderer6343.multiblock.structure=Multiblock Structure
+blockrenderer6343.no_space=Not enough space!
+blockrenderer6343.multiblock.project=Project Multiblock
+blockrenderer6343.multiblock.overlay=NEI Overlay
